Redesigned the CARLA-side ZMQ communication architecture
